MPP-1921: Add ./manage.py send_simulator_email
Add management command send_simulator_email to send an email from a
Relay user to an AWS SES simulator mailbox. These mailboxes have
set responses that include soft bounces, out-of-office replies, and
marking as spam.

The new code and some existing code is in a new file emails/ses.py, to
centralize the Relay interface to SES. Both this file and the management
command are fully tested.