MPP-1921: Reimplement SES types and serializers
Move the SES types to emails/ses_types.py, and remove the validation
functions. This means that ses_types.py is just a bunch of class
declarations (100% coverage!).

Re-implement the validation logic as Django REST Framework serializers
in ses_serializers.py. This is mostly a direct translation of
ses_types.py, with some quirks. This avoids magic like ModelSerializers,
which could reduce the duplication, but DRF upgrades tend to break
magic.

This adds a _lot_ of fixtures, with the extension `_ses_body.json`.
These were used to test the new serializers and types.  Many of these
are copied from AWS documentation, which is 95% accurate, and from live
responses from the AWS SES simulator mailboxes, which feels more
accurate.

Two existing fixtures were updated to give them valid values. The soft
bounce fixture's subtype was changed from "SRETeamEatenByDinosaurs" to
"MailboxFull", and the spam bounce from "StopRelayingSpamForThisUser" to
"ContentRejected".

A README.md is added to the fixtures to document their sources and
modifications.

After removing types from `emails/ses.py`, it is a lot smaller and
mostly functions. The tests have been reduced as well, since processing
the fixtures covers almost all of the code.

Minimal changes are made to `emails/views.py`. Bounce and Recieved
messages _are_ processed and validated by the new code, but the code
continues to use the dict structure loaded from JSON.